+++ src/main/java/eterea/tenant/service/tenant/hexagonal/articulomovimiento/domain/model/ArticuloMovimiento.java
@@ -2,6 +2,7 @@ package eterea.tenant.service.tenant.hexagonal.articulomovimiento.domain.model;
 
 import eterea.tenant.service.tenant.hexagonal.articulo.domain.model.Articulo;
 import eterea.tenant.service.tool.Jsonifier;
+import eterea.tenant.service.tool.Jsonifyable;
 import lombok.*;
 
 import java.math.BigDecimal;
@@ -12,7 +13,7 @@ import java.time.OffsetDateTime;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class ArticuloMovimiento {
+public class ArticuloMovimiento implements Jsonifyable {
 
     private Long articuloMovimientoId;
     private Long clienteMovimientoId;
@@ -44,7 +45,4 @@ public class ArticuloMovimiento {
     private BigDecimal totalSinIva;
     private Articulo articulo;
 
-    public String jsonify() {
-        return Jsonifier.builder(this).build();
-    }
 }