+++ src/main/java/eterea/tenant/service/tenant/hexagonal/articulomovimiento/infrastructure/persistence/entity/ArticuloMovimientoEntity.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 import eterea.tenant.service.tenant.hexagonal.articulo.infrastructure.persistence.entity.ArticuloEntity;
 import eterea.tenant.service.tenant.model.Auditable;
 import eterea.tenant.service.tool.Jsonifier;
+import eterea.tenant.service.tool.Jsonifyable;
 import jakarta.persistence.*;
 import lombok.*;
 
@@ -17,7 +18,7 @@ import java.time.OffsetDateTime;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class ArticuloMovimientoEntity extends Auditable {
+public class ArticuloMovimientoEntity extends Auditable implements Jsonifyable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -125,9 +126,5 @@ public class ArticuloMovimientoEntity extends Auditable {
     @OneToOne(optional = true)
     @JoinColumn(name = "cgoartic", updatable = false, insertable = false)
     private ArticuloEntity articulo;
-
-    public String jsonify() {
-        return Jsonifier.builder(this).build();
-    }
 
 }