+++ src/main/java/eterea/tenant/service/tenant/hexagonal/invoicedata/infrastructure/dto/InvoiceDataResponse.java
@@ -1,18 +1,15 @@
 package eterea.tenant.service.tenant.hexagonal.invoicedata.infrastructure.dto;
 
 import eterea.tenant.service.tool.Jsonifier;
+import eterea.tenant.service.tool.Jsonifyable;
 import lombok.*;
 
 @Getter
 @Builder
-public class InvoiceDataResponse {
+public class InvoiceDataResponse implements Jsonifyable {
 
     ClienteMovimientoResponse clienteMovimiento;
     RegistroCaeResponse registroCae;
     ClienteMovimientoResponse clienteMovimientoAsociado;
-
-    public String jsonify() {
-        return Jsonifier.builder(this).build();
-    }
 
 }