+++ src/main/java/eterea/tenant/service/tenant/hexagonal/invoicedata/infrastructure/dto/RegistroCaeResponse.java
@@ -1,6 +1,7 @@
 package eterea.tenant.service.tenant.hexagonal.invoicedata.infrastructure.dto;
 
 import eterea.tenant.service.tool.Jsonifier;
+import eterea.tenant.service.tool.Jsonifyable;
 import lombok.Builder;
 import lombok.Getter;
 
@@ -8,7 +9,7 @@ import java.math.BigDecimal;
 
 @Getter
 @Builder
-public class RegistroCaeResponse {
+public class RegistroCaeResponse implements Jsonifyable {
 
     private Integer tipoDocumento;
     private Integer puntoVenta;
@@ -20,9 +21,5 @@ public class RegistroCaeResponse {
     private String caeVencimiento;
     private String fecha;
     private ComprobanteResponse comprobante;
-
-    public String jsonify() {
-        return Jsonifier.builder(this).build();
-    }
 
 }