+++ src/main/java/eterea/tenant/service/tenant/model/dto/TransferenciaWrapperDto.java
@@ -3,6 +3,7 @@ package eterea.tenant.service.tenant.model.dto;
 import eterea.tenant.service.tenant.kotlin.model.ValorMovimiento;
 import eterea.tenant.service.tenant.model.CuentaMovimiento;
 import eterea.tenant.service.tool.Jsonifier;
+import eterea.tenant.service.tool.Jsonifyable;
 import lombok.*;
 
 import java.util.List;
@@ -12,14 +13,10 @@ import java.util.List;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class TransferenciaWrapperDto {
+public class TransferenciaWrapperDto implements Jsonifyable {
 
     private TransferenciaDto transferencia;
     private List<ValorMovimiento> valorMovimientos;
     private List<CuentaMovimiento> cuentaMovimientos;
-
-    public String jsonify() {
-        return Jsonifier.builder(this).build();
-    }
 
 }